+++ src/test/java/org/opensearch/neuralsearch/query/HybridQueryIT.java
@@ -28,7 +28,9 @@ import org.opensearch.client.ResponseException;
 import org.opensearch.index.query.BoolQueryBuilder;
 import org.opensearch.index.query.MatchQueryBuilder;
 import org.opensearch.index.query.NestedQueryBuilder;
+import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.index.query.QueryBuilders;
+import org.opensearch.index.query.RangeQueryBuilder;
 import org.opensearch.index.query.TermQueryBuilder;
 import org.opensearch.neuralsearch.BaseNeuralSearchIT;
 
@@ -39,10 +41,13 @@ import lombok.SneakyThrows;
 public class HybridQueryIT extends BaseNeuralSearchIT {
     private static final String TEST_BASIC_INDEX_NAME = "test-hybrid-basic-index";
     private static final String TEST_BASIC_VECTOR_DOC_FIELD_INDEX_NAME = "test-hybrid-vector-doc-field-index";
+    private static final String TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME = "test-hybrid-multi-doc-nested-fields-index";
     private static final String TEST_MULTI_DOC_INDEX_NAME = "test-hybrid-multi-doc-index";
     private static final String TEST_MULTI_DOC_INDEX_NAME_ONE_SHARD = "test-hybrid-multi-doc-single-shard-index";
     private static final String TEST_MULTI_DOC_INDEX_WITH_NESTED_TYPE_NAME_ONE_SHARD =
         "test-hybrid-multi-doc-nested-type-single-shard-index";
+    private static final String TEST_INDEX_WITH_KEYWORDS_ONE_SHARD = "test-hybrid-keywords-single-shard-index";
+    private static final String TEST_INDEX_WITH_KEYWORDS_THREE_SHARDS = "test-hybrid-keywords-three-shards-index";
     private static final String TEST_QUERY_TEXT = "greetings";
     private static final String TEST_QUERY_TEXT2 = "salute";
     private static final String TEST_QUERY_TEXT3 = "hello";
@@ -59,6 +64,18 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
     private static final String NESTED_FIELD_2 = "lastname";
     private static final String NESTED_FIELD_1_VALUE = "john";
     private static final String NESTED_FIELD_2_VALUE = "black";
+    private static final String KEYWORD_FIELD_1 = "doc_keyword";
+    private static final String KEYWORD_FIELD_1_VALUE = "workable";
+    private static final String KEYWORD_FIELD_2_VALUE = "angry";
+    private static final String KEYWORD_FIELD_3_VALUE = "likeable";
+    private static final String KEYWORD_FIELD_4_VALUE = "entire";
+    private static final String INTEGER_FIELD_PRICE = "doc_price";
+    private static final int INTEGER_FIELD_PRICE_1_VALUE = 130;
+    private static final int INTEGER_FIELD_PRICE_2_VALUE = 100;
+    private static final int INTEGER_FIELD_PRICE_3_VALUE = 200;
+    private static final int INTEGER_FIELD_PRICE_4_VALUE = 25;
+    private static final int INTEGER_FIELD_PRICE_5_VALUE = 30;
+    private static final int INTEGER_FIELD_PRICE_6_VALUE = 350;
     private final float[] testVector1 = createRandomVector(TEST_DIMENSION);
     private final float[] testVector2 = createRandomVector(TEST_DIMENSION);
     private final float[] testVector3 = createRandomVector(TEST_DIMENSION);
@@ -114,10 +131,8 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
      */
     @SneakyThrows
     public void testComplexQuery_whenMultipleSubqueries_thenSuccessful() {
-        String modelId = null;
         try {
             initializeIndexIfNotExist(TEST_BASIC_VECTOR_DOC_FIELD_INDEX_NAME);
-            modelId = prepareModel();
             createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
             TermQueryBuilder termQueryBuilder1 = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT3);
             TermQueryBuilder termQueryBuilder2 = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT4);
@@ -158,7 +173,7 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
             assertNotNull(total.get("relation"));
             assertEquals(RELATION_EQUAL_TO, total.get("relation"));
         } finally {
-            wipeOfTestResources(TEST_BASIC_VECTOR_DOC_FIELD_INDEX_NAME, null, modelId, SEARCH_PIPELINE);
+            wipeOfTestResources(TEST_BASIC_VECTOR_DOC_FIELD_INDEX_NAME, null, null, SEARCH_PIPELINE);
         }
     }
 
@@ -190,10 +205,8 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
      */
     @SneakyThrows
     public void testComplexQuery_whenMultipleIdenticalSubQueries_thenSuccessful() {
-        String modelId = null;
         try {
             initializeIndexIfNotExist(TEST_BASIC_VECTOR_DOC_FIELD_INDEX_NAME);
-            modelId = prepareModel();
             createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
             TermQueryBuilder termQueryBuilder1 = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT3);
             TermQueryBuilder termQueryBuilder2 = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT4);
@@ -233,16 +246,14 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
             assertNotNull(total.get("relation"));
             assertEquals(RELATION_EQUAL_TO, total.get("relation"));
         } finally {
-            wipeOfTestResources(TEST_BASIC_VECTOR_DOC_FIELD_INDEX_NAME, null, modelId, SEARCH_PIPELINE);
+            wipeOfTestResources(TEST_BASIC_VECTOR_DOC_FIELD_INDEX_NAME, null, null, SEARCH_PIPELINE);
         }
     }
 
     @SneakyThrows
     public void testNoMatchResults_whenOnlyTermSubQueryWithoutMatch_thenEmptyResult() {
-        String modelId = null;
         try {
-            initializeIndexIfNotExist(TEST_MULTI_DOC_INDEX_NAME);
-            modelId = prepareModel();
+            initializeIndexIfNotExist(TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME);
             createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
             TermQueryBuilder termQueryBuilder = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT);
             TermQueryBuilder termQuery2Builder = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT2);
@@ -251,7 +262,7 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
             hybridQueryBuilderOnlyTerm.add(termQuery2Builder);
 
             Map<String, Object> searchResponseAsMap = search(
-                TEST_MULTI_DOC_INDEX_NAME,
+                TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME,
                 hybridQueryBuilderOnlyTerm,
                 null,
                 10,
@@ -268,16 +279,14 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
             assertNotNull(total.get("relation"));
             assertEquals(RELATION_EQUAL_TO, total.get("relation"));
         } finally {
-            wipeOfTestResources(TEST_MULTI_DOC_INDEX_NAME, null, modelId, SEARCH_PIPELINE);
+            wipeOfTestResources(TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME, null, null, SEARCH_PIPELINE);
         }
     }
 
     @SneakyThrows
     public void testNestedQuery_whenHybridQueryIsWrappedIntoOtherQuery_thenFail() {
-        String modelId = null;
         try {
             initializeIndexIfNotExist(TEST_MULTI_DOC_INDEX_NAME_ONE_SHARD);
-            modelId = prepareModel();
             createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
             MatchQueryBuilder matchQueryBuilder = QueryBuilders.matchQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT3);
             MatchQueryBuilder matchQuery2Builder = QueryBuilders.matchQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT4);
@@ -321,16 +330,14 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
                 )
             );
         } finally {
-            wipeOfTestResources(TEST_MULTI_DOC_INDEX_NAME_ONE_SHARD, null, modelId, SEARCH_PIPELINE);
+            wipeOfTestResources(TEST_MULTI_DOC_INDEX_NAME_ONE_SHARD, null, null, SEARCH_PIPELINE);
         }
     }
 
     @SneakyThrows
     public void testIndexWithNestedFields_whenHybridQuery_thenSuccess() {
-        String modelId = null;
         try {
             initializeIndexIfNotExist(TEST_MULTI_DOC_INDEX_WITH_NESTED_TYPE_NAME_ONE_SHARD);
-            modelId = prepareModel();
             createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
             TermQueryBuilder termQueryBuilder = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT3);
             TermQueryBuilder termQuery2Builder = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT2);
@@ -356,16 +363,14 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
             assertNotNull(total.get("relation"));
             assertEquals(RELATION_EQUAL_TO, total.get("relation"));
         } finally {
-            wipeOfTestResources(TEST_MULTI_DOC_INDEX_WITH_NESTED_TYPE_NAME_ONE_SHARD, null, modelId, SEARCH_PIPELINE);
+            wipeOfTestResources(TEST_MULTI_DOC_INDEX_WITH_NESTED_TYPE_NAME_ONE_SHARD, null, null, SEARCH_PIPELINE);
         }
     }
 
     @SneakyThrows
     public void testIndexWithNestedFields_whenHybridQueryIncludesNested_thenSuccess() {
-        String modelId = null;
         try {
             initializeIndexIfNotExist(TEST_MULTI_DOC_INDEX_WITH_NESTED_TYPE_NAME_ONE_SHARD);
-            modelId = prepareModel();
             createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
             TermQueryBuilder termQueryBuilder = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT);
             NestedQueryBuilder nestedQueryBuilder = QueryBuilders.nestedQuery(
@@ -395,7 +400,263 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
             assertNotNull(total.get("relation"));
             assertEquals(RELATION_EQUAL_TO, total.get("relation"));
         } finally {
-            wipeOfTestResources(TEST_MULTI_DOC_INDEX_WITH_NESTED_TYPE_NAME_ONE_SHARD, null, modelId, SEARCH_PIPELINE);
+            wipeOfTestResources(TEST_MULTI_DOC_INDEX_WITH_NESTED_TYPE_NAME_ONE_SHARD, null, null, SEARCH_PIPELINE);
+        }
+    }
+
+    @SneakyThrows
+    public void testRequestCache_whenOneShardAndQueryReturnResults_thenSuccessful() {
+        try {
+            initializeIndexIfNotExist(TEST_INDEX_WITH_KEYWORDS_ONE_SHARD);
+            createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
+            MatchQueryBuilder matchQueryBuilder = QueryBuilders.matchQuery(KEYWORD_FIELD_1, KEYWORD_FIELD_2_VALUE);
+            RangeQueryBuilder rangeQueryBuilder = QueryBuilders.rangeQuery(INTEGER_FIELD_PRICE).gte(10).lte(1000);
+
+            HybridQueryBuilder hybridQueryBuilder = new HybridQueryBuilder();
+            hybridQueryBuilder.add(matchQueryBuilder);
+            hybridQueryBuilder.add(rangeQueryBuilder);
+
+            // first query with cache flag executed normally by reading documents from index
+            Map<String, Object> firstSearchResponseAsMap = search(
+                TEST_INDEX_WITH_KEYWORDS_ONE_SHARD,
+                hybridQueryBuilder,
+                null,
+                10,
+                Map.of("search_pipeline", SEARCH_PIPELINE, "request_cache", Boolean.TRUE.toString())
+            );
+
+            int firstQueryHitCount = getHitCount(firstSearchResponseAsMap);
+            assertTrue(firstQueryHitCount > 0);
+
+            List<Map<String, Object>> hitsNestedList = getNestedHits(firstSearchResponseAsMap);
+            List<String> ids = new ArrayList<>();
+            List<Double> scores = new ArrayList<>();
+            for (Map<String, Object> oneHit : hitsNestedList) {
+                ids.add((String) oneHit.get("_id"));
+                scores.add((Double) oneHit.get("_score"));
+            }
+
+            // verify that scores are in desc order
+            assertTrue(IntStream.range(0, scores.size() - 1).noneMatch(idx -> scores.get(idx) < scores.get(idx + 1)));
+            // verify that all ids are unique
+            assertEquals(Set.copyOf(ids).size(), ids.size());
+
+            Map<String, Object> total = getTotalHits(firstSearchResponseAsMap);
+            assertNotNull(total.get("value"));
+            assertEquals(firstQueryHitCount, total.get("value"));
+            assertNotNull(total.get("relation"));
+            assertEquals(RELATION_EQUAL_TO, total.get("relation"));
+
+            // second query is served from the cache
+            Map<String, Object> secondSearchResponseAsMap = search(
+                TEST_INDEX_WITH_KEYWORDS_ONE_SHARD,
+                hybridQueryBuilder,
+                null,
+                10,
+                Map.of("search_pipeline", SEARCH_PIPELINE, "request_cache", Boolean.TRUE.toString())
+            );
+
+            assertEquals(firstQueryHitCount, getHitCount(secondSearchResponseAsMap));
+
+            List<Map<String, Object>> hitsNestedListSecondQuery = getNestedHits(secondSearchResponseAsMap);
+            List<String> idsSecondQuery = new ArrayList<>();
+            List<Double> scoresSecondQuery = new ArrayList<>();
+            for (Map<String, Object> oneHit : hitsNestedListSecondQuery) {
+                idsSecondQuery.add((String) oneHit.get("_id"));
+                scoresSecondQuery.add((Double) oneHit.get("_score"));
+            }
+
+            // verify that scores are in desc order
+            assertTrue(
+                IntStream.range(0, scoresSecondQuery.size() - 1)
+                    .noneMatch(idx -> scoresSecondQuery.get(idx) < scoresSecondQuery.get(idx + 1))
+            );
+            // verify that all ids are unique
+            assertEquals(Set.copyOf(idsSecondQuery).size(), idsSecondQuery.size());
+
+            Map<String, Object> totalSecondQuery = getTotalHits(secondSearchResponseAsMap);
+            assertNotNull(totalSecondQuery.get("value"));
+            assertEquals(firstQueryHitCount, totalSecondQuery.get("value"));
+            assertNotNull(totalSecondQuery.get("relation"));
+            assertEquals(RELATION_EQUAL_TO, totalSecondQuery.get("relation"));
+        } finally {
+            wipeOfTestResources(TEST_INDEX_WITH_KEYWORDS_ONE_SHARD, null, null, SEARCH_PIPELINE);
+        }
+    }
+
+    @SneakyThrows
+    public void testRequestCache_whenMultipleShardsQueryReturnResults_thenSuccessful() {
+        try {
+            initializeIndexIfNotExist(TEST_INDEX_WITH_KEYWORDS_THREE_SHARDS);
+            createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
+            MatchQueryBuilder matchQueryBuilder = QueryBuilders.matchQuery(KEYWORD_FIELD_1, KEYWORD_FIELD_2_VALUE);
+            RangeQueryBuilder rangeQueryBuilder = QueryBuilders.rangeQuery(INTEGER_FIELD_PRICE).gte(10).lte(1000);
+
+            HybridQueryBuilder hybridQueryBuilder = new HybridQueryBuilder();
+            hybridQueryBuilder.add(matchQueryBuilder);
+            hybridQueryBuilder.add(rangeQueryBuilder);
+
+            // first query with cache flag executed normally by reading documents from index
+            Map<String, Object> firstSearchResponseAsMap = search(
+                TEST_INDEX_WITH_KEYWORDS_THREE_SHARDS,
+                hybridQueryBuilder,
+                null,
+                10,
+                Map.of("search_pipeline", SEARCH_PIPELINE, "request_cache", Boolean.TRUE.toString())
+            );
+
+            int firstQueryHitCount = getHitCount(firstSearchResponseAsMap);
+            assertTrue(firstQueryHitCount > 0);
+
+            List<Map<String, Object>> hitsNestedList = getNestedHits(firstSearchResponseAsMap);
+            List<String> ids = new ArrayList<>();
+            List<Double> scores = new ArrayList<>();
+            for (Map<String, Object> oneHit : hitsNestedList) {
+                ids.add((String) oneHit.get("_id"));
+                scores.add((Double) oneHit.get("_score"));
+            }
+
+            // verify that scores are in desc order
+            assertTrue(IntStream.range(0, scores.size() - 1).noneMatch(idx -> scores.get(idx) < scores.get(idx + 1)));
+            // verify that all ids are unique
+            assertEquals(Set.copyOf(ids).size(), ids.size());
+
+            Map<String, Object> total = getTotalHits(firstSearchResponseAsMap);
+            assertNotNull(total.get("value"));
+            assertEquals(firstQueryHitCount, total.get("value"));
+            assertNotNull(total.get("relation"));
+            assertEquals(RELATION_EQUAL_TO, total.get("relation"));
+
+            // second query is served from the cache
+            Map<String, Object> secondSearchResponseAsMap = search(
+                TEST_INDEX_WITH_KEYWORDS_THREE_SHARDS,
+                hybridQueryBuilder,
+                null,
+                10,
+                Map.of("search_pipeline", SEARCH_PIPELINE, "request_cache", Boolean.TRUE.toString())
+            );
+
+            assertEquals(firstQueryHitCount, getHitCount(secondSearchResponseAsMap));
+
+            List<Map<String, Object>> hitsNestedListSecondQuery = getNestedHits(secondSearchResponseAsMap);
+            List<String> idsSecondQuery = new ArrayList<>();
+            List<Double> scoresSecondQuery = new ArrayList<>();
+            for (Map<String, Object> oneHit : hitsNestedListSecondQuery) {
+                idsSecondQuery.add((String) oneHit.get("_id"));
+                scoresSecondQuery.add((Double) oneHit.get("_score"));
+            }
+
+            // verify that scores are in desc order
+            assertTrue(
+                IntStream.range(0, scoresSecondQuery.size() - 1)
+                    .noneMatch(idx -> scoresSecondQuery.get(idx) < scoresSecondQuery.get(idx + 1))
+            );
+            // verify that all ids are unique
+            assertEquals(Set.copyOf(idsSecondQuery).size(), idsSecondQuery.size());
+
+            Map<String, Object> totalSecondQuery = getTotalHits(secondSearchResponseAsMap);
+            assertNotNull(totalSecondQuery.get("value"));
+            assertEquals(firstQueryHitCount, totalSecondQuery.get("value"));
+            assertNotNull(totalSecondQuery.get("relation"));
+            assertEquals(RELATION_EQUAL_TO, totalSecondQuery.get("relation"));
+        } finally {
+            wipeOfTestResources(TEST_INDEX_WITH_KEYWORDS_THREE_SHARDS, null, null, SEARCH_PIPELINE);
+        }
+    }
+
+    @SneakyThrows
+    public void testWrappedQueryWithFilter_whenIndexAliasHasFilterAndIndexWithNestedFields_thenSuccess() {
+        String modelId = null;
+        String alias = "alias_with_filter";
+        try {
+            initializeIndexIfNotExist(TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME);
+            modelId = prepareModel();
+            createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
+            // create alias for index
+            QueryBuilder aliasFilter = QueryBuilders.boolQuery()
+                .mustNot(QueryBuilders.matchQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT3));
+            createIndexAlias(TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME, alias, aliasFilter);
+
+            NeuralQueryBuilder neuralQueryBuilder = new NeuralQueryBuilder(
+                TEST_KNN_VECTOR_FIELD_NAME_1,
+                TEST_QUERY_TEXT,
+                "",
+                modelId,
+                5,
+                null,
+                null
+            );
+            HybridQueryBuilder hybridQueryBuilder = new HybridQueryBuilder();
+            hybridQueryBuilder.add(neuralQueryBuilder);
+
+            Map<String, Object> searchResponseAsMap = search(
+                alias,
+                hybridQueryBuilder,
+                null,
+                10,
+                Map.of("search_pipeline", SEARCH_PIPELINE)
+            );
+
+            assertEquals(2, getHitCount(searchResponseAsMap));
+            assertTrue(getMaxScore(searchResponseAsMap).isPresent());
+            assertEquals(1.0f, getMaxScore(searchResponseAsMap).get(), DELTA_FOR_SCORE_ASSERTION);
+
+            Map<String, Object> total = getTotalHits(searchResponseAsMap);
+            assertNotNull(total.get("value"));
+            assertEquals(2, total.get("value"));
+            assertNotNull(total.get("relation"));
+            assertEquals(RELATION_EQUAL_TO, total.get("relation"));
+        } finally {
+            deleteIndexAlias(TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME, alias);
+            wipeOfTestResources(TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME, null, modelId, SEARCH_PIPELINE);
+        }
+    }
+
+    @SneakyThrows
+    public void testWrappedQueryWithFilter_whenIndexAliasHasFilters_thenSuccess() {
+        String modelId = null;
+        String alias = "alias_with_filter";
+        try {
+            initializeIndexIfNotExist(TEST_MULTI_DOC_INDEX_NAME);
+            modelId = prepareModel();
+            createSearchPipelineWithResultsPostProcessor(SEARCH_PIPELINE);
+            // create alias for index
+            QueryBuilder aliasFilter = QueryBuilders.boolQuery()
+                .mustNot(QueryBuilders.matchQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT3));
+            createIndexAlias(TEST_MULTI_DOC_INDEX_NAME, alias, aliasFilter);
+
+            NeuralQueryBuilder neuralQueryBuilder = new NeuralQueryBuilder(
+                TEST_KNN_VECTOR_FIELD_NAME_1,
+                TEST_QUERY_TEXT,
+                "",
+                modelId,
+                5,
+                null,
+                null
+            );
+            HybridQueryBuilder hybridQueryBuilder = new HybridQueryBuilder();
+            hybridQueryBuilder.add(neuralQueryBuilder);
+
+            Map<String, Object> searchResponseAsMap = search(
+                alias,
+                hybridQueryBuilder,
+                null,
+                10,
+                Map.of("search_pipeline", SEARCH_PIPELINE)
+            );
+
+            assertEquals(2, getHitCount(searchResponseAsMap));
+            assertTrue(getMaxScore(searchResponseAsMap).isPresent());
+            assertEquals(1.0f, getMaxScore(searchResponseAsMap).get(), DELTA_FOR_SCORE_ASSERTION);
+
+            Map<String, Object> total = getTotalHits(searchResponseAsMap);
+            assertNotNull(total.get("value"));
+            assertEquals(2, total.get("value"));
+            assertNotNull(total.get("relation"));
+            assertEquals(RELATION_EQUAL_TO, total.get("relation"));
+        } finally {
+            deleteIndexAlias(TEST_MULTI_DOC_INDEX_NAME, alias);
+            wipeOfTestResources(TEST_MULTI_DOC_INDEX_NAME, null, modelId, SEARCH_PIPELINE);
         }
     }
 
@@ -449,10 +710,28 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
             assertEquals(3, getDocCount(TEST_BASIC_VECTOR_DOC_FIELD_INDEX_NAME));
         }
 
+        if (TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME.equals(indexName) && !indexExists(TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME)) {
+            createIndexWithConfiguration(
+                indexName,
+                buildIndexConfiguration(
+                    Collections.singletonList(new KNNFieldConfig(TEST_KNN_VECTOR_FIELD_NAME_1, TEST_DIMENSION, TEST_SPACE_TYPE)),
+                    List.of(TEST_NESTED_TYPE_FIELD_NAME_1),
+                    1
+                ),
+                ""
+            );
+            addDocsToIndex(TEST_MULTI_DOC_WITH_NESTED_FIELDS_INDEX_NAME);
+        }
+
         if (TEST_MULTI_DOC_INDEX_NAME.equals(indexName) && !indexExists(TEST_MULTI_DOC_INDEX_NAME)) {
-            prepareKnnIndex(
-                TEST_MULTI_DOC_INDEX_NAME,
-                Collections.singletonList(new KNNFieldConfig(TEST_KNN_VECTOR_FIELD_NAME_1, TEST_DIMENSION, TEST_SPACE_TYPE))
+            createIndexWithConfiguration(
+                indexName,
+                buildIndexConfiguration(
+                    Collections.singletonList(new KNNFieldConfig(TEST_KNN_VECTOR_FIELD_NAME_1, TEST_DIMENSION, TEST_SPACE_TYPE)),
+                    List.of(),
+                    1
+                ),
+                ""
             );
             addDocsToIndex(TEST_MULTI_DOC_INDEX_NAME);
         }
@@ -488,6 +767,104 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
                 List.of(),
                 List.of(TEST_NESTED_TYPE_FIELD_NAME_1),
                 List.of(Map.of(NESTED_FIELD_1, NESTED_FIELD_1_VALUE, NESTED_FIELD_2, NESTED_FIELD_2_VALUE))
+            );
+        }
+
+        if (TEST_INDEX_WITH_KEYWORDS_ONE_SHARD.equals(indexName) && !indexExists(TEST_INDEX_WITH_KEYWORDS_ONE_SHARD)) {
+            createIndexWithConfiguration(
+                indexName,
+                buildIndexConfiguration(List.of(), List.of(), List.of(INTEGER_FIELD_PRICE), List.of(KEYWORD_FIELD_1), List.of(), 1),
+                ""
+            );
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "1",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_1_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_1_VALUE
+            );
+            addDocWithKeywordsAndIntFields(indexName, "2", INTEGER_FIELD_PRICE, INTEGER_FIELD_PRICE_2_VALUE, null, null);
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "3",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_3_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_2_VALUE
+            );
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "4",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_4_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_3_VALUE
+            );
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "5",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_5_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_4_VALUE
+            );
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "6",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_6_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_4_VALUE
+            );
+        }
+
+        if (TEST_INDEX_WITH_KEYWORDS_THREE_SHARDS.equals(indexName) && !indexExists(TEST_INDEX_WITH_KEYWORDS_THREE_SHARDS)) {
+            createIndexWithConfiguration(
+                indexName,
+                buildIndexConfiguration(List.of(), List.of(), List.of(INTEGER_FIELD_PRICE), List.of(KEYWORD_FIELD_1), List.of(), 3),
+                ""
+            );
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "1",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_1_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_1_VALUE
+            );
+            addDocWithKeywordsAndIntFields(indexName, "2", INTEGER_FIELD_PRICE, INTEGER_FIELD_PRICE_2_VALUE, null, null);
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "3",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_3_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_2_VALUE
+            );
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "4",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_4_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_3_VALUE
+            );
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "5",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_5_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_4_VALUE
+            );
+            addDocWithKeywordsAndIntFields(
+                indexName,
+                "6",
+                INTEGER_FIELD_PRICE,
+                INTEGER_FIELD_PRICE_6_VALUE,
+                KEYWORD_FIELD_1,
+                KEYWORD_FIELD_4_VALUE
             );
         }
     }
@@ -531,5 +908,36 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
     private Optional<Float> getMaxScore(Map<String, Object> searchResponseAsMap) {
         Map<String, Object> hitsMap = (Map<String, Object>) searchResponseAsMap.get("hits");
         return hitsMap.get("max_score") == null ? Optional.empty() : Optional.of(((Double) hitsMap.get("max_score")).floatValue());
+    }
+
+    private void addDocWithKeywordsAndIntFields(
+        final String indexName,
+        final String docId,
+        final String integerField,
+        final Integer integerFieldValue,
+        final String keywordField,
+        final String keywordFieldValue
+    ) {
+        List<String> intFields = integerField == null ? List.of() : List.of(integerField);
+        List<Integer> intValues = integerFieldValue == null ? List.of() : List.of(integerFieldValue);
+        List<String> keywordFields = keywordField == null ? List.of() : List.of(keywordField);
+        List<String> keywordValues = keywordFieldValue == null ? List.of() : List.of(keywordFieldValue);
+
+        addKnnDoc(
+            indexName,
+            docId,
+            List.of(),
+            List.of(),
+            List.of(),
+            List.of(),
+            List.of(),
+            List.of(),
+            intFields,
+            intValues,
+            keywordFields,
+            keywordValues,
+            List.of(),
+            List.of()
+        );
     }
 }